+++ examples/stream-example/README.md
@@ -1,0 +1,17 @@
+## run producer
+
+```sh
+cargo run --bin producer
+```
+
+access endpoint send msg:
+
+```sh
+curl http://localhost:8080
+```
+
+## run consumer
+
+```sh
+cargo run --bin consumer
+```
