+++ examples/stream-example/src/lib.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct Payload {
+    success: bool,
+    msg: String,
+}
