+++ spring-boot/src/lib.rs
@@ -4,3 +4,4 @@ pub mod error;
 pub mod log;
 pub mod plugin;
 pub use async_trait::async_trait;
+pub use tracing;
