+++ spring/src/lib.rs
@@ -1,2 +1,3 @@
 pub use spring_boot::app::App;
+pub use spring_boot::tracing;
 pub use spring_macros::*;
